+++ BootcampMagic/Controller/ProjectTabBarController.swift
@@ -8,11 +8,11 @@
 import UIKit
 
 class ProjectTabBarController: UITabBarController {
-    let projectTBC = UITabBarController()
+    let projectTabBarController = UITabBarController()
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        creatTBC()
+        setupTabBar()
         assignBackground()
     }
 
@@ -33,22 +33,22 @@ class ProjectTabBarController: UITabBarController {
         self.view.sendSubviewToBack(imageView)
       }
 
-    func creatTBC() {
+    func setupTabBar() {
         UITabBarItem.appearance().setTitleTextAttributes([NSAttributedString.Key.font: TextFont.button], for: .normal)
         UITabBarItem.appearance().setTitleTextAttributes([NSAttributedString.Key.font: TextFont.button], for: .selected)
 
         UITabBar.appearance().tintColor = TextColor.button
         UITabBar.setTransparentTabbar()
-        viewControllers = [creatEVC(), creatFVC()]
+        viewControllers = [setupExpansionsViewController(), setupFavoritesViewController()]
     }
 
-    func creatEVC() -> UINavigationController {
+    func setupExpansionsViewController() -> UINavigationController {
         let firstViewController = UINavigationController(rootViewController: ExpansionsViewController())
         firstViewController.title = "Expansions"
         return firstViewController
     }
 
-    func creatFVC() -> UINavigationController {
+    func setupFavoritesViewController() -> UINavigationController {
         let secondViewController = UINavigationController(rootViewController: FavoritesViewController())
         secondViewController.title = "Favorites"
         return secondViewController
